+++ packages/plugin-analytics-api/src/analytics-plugin.module.ts
@@ -1,7 +1,8 @@
 import { Module } from "@nestjs/common";
 import { AnalyticsModule } from "./analytics";
+import { SettingsModule } from "@kitejs-cms/core";
 
 @Module({
-  imports: [AnalyticsModule],
+  imports: [AnalyticsModule, SettingsModule],
 })
 export class AnalyticsPluginModule {}
