+++ packages/plugin-analytics-api/src/analytics/models/analytics-plugin-settings.model.ts
@@ -1,0 +1,5 @@
+export type AnalyticsPluginSettingsModel = {
+  apiKey: string;
+  /** Number of days to retain analytics data before pruning */
+  retentionDays: number;
+};
