+++ packages/plugin-analytics-api/src/constants.ts
@@ -1,4 +1,5 @@
 import { SettingModel } from "@kitejs-cms/core";
+import { AnalyticsPluginSettingsModel } from "./analytics/models/analytics-plugin-settings.model";
 import { randomUUID } from "crypto";
 
 export const ANALYTICS_PLUGIN_NAMESPACE = "analytics";
@@ -6,12 +7,6 @@ export const ANALYTICS_PLUGIN_NAMESPACE = "analytics";
 export const ANALYTICS_SETTINGS_KEY = `${ANALYTICS_PLUGIN_NAMESPACE}:config`;
 
 export const DEFAULT_RETENTION_DAYS = 90;
-
-export type AnalyticsPluginSettingsModel = {
-  apiKey: string;
-  /** Number of days to retain analytics data before pruning */
-  retentionDays: number;
-};
 
 export const AnalyticsPermissions = [];
